featrue: SEQUOIADBMAINSTREAM-3079 民生资源管控项目代码移值 增加 节点调度代码
git-svn-id: http://192.168.20.11/sequoiadb/trunk@35146 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/coord/coordCommandStat.cpp
@@ -410,7 +410,7 @@ namespace engine
                                                rtnContext *pContext,
                                                pmdEDUCB *cb )
    {
-      return pContext->append( BSON( FIELD_NAME_TOTAL << 1 ) ) ;
+      return pContext->append( BSON( FIELD_NAME_TOTAL << (INT64)1 ) ) ;
    }
 
    /*
